refactor: Move enum to individual file
